third_party/WebKit/LayoutTests/http/tests/websocket/tests/hybi/close-on-navigate-new-location.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="/js-test-resources/js-test-pre.js"></script>
</head>
<body>
<div id="description"></div>
<div id="console"></div>
<script type="text/javascript">
description("Test if Web Socket is closed when navigating to new location.");

window.jsTestIsAsync = true;

var frameDiv;
var closedSocket;
var closeCode;

var ws = new WebSocket("ws://127.0.0.1:8880/websocket/tests/hybi/close-on-unload");
ws.onopen = function()
{
    debug("PASS ws on master document is ready.");
    frameDiv = document.createElement("iframe");
    frameDiv.src = "resources/close-on-unload-iframe.html";
    document.body.appendChild(frameDiv);
    debug("PASS insert a iframe, where open ws called 'socket1'");
};
ws.onmessage = function(evt)
{
    var messageData = JSON.parse(evt.data);
    closedSocket = messageData.message;
    closeCode = messageData.closeCode;
    ws.close();
};
ws.onclose = function()
{
    shouldBe("closedSocket", '"socket1"');
    shouldBe("closeCode", '"1001"');
    finishJSTest();
};

document.iframeReady = function()
{
    debug("PASS 'socket1' is sent to the server. navigate to new location. expect receiving 'socket1' on ws...");
    frameDiv.src = "../handler_map.txt";
};

</script>
<script src="/js-test-resources/js-test-post.js"></script>
</body>
</html>
